fix index file name and location

+++ src/benchmark/generator.rs
@@ -2,7 +2,7 @@ use rand::{rngs::*, RngCore};
 
 pub struct Generator {
     config: Config,
-    written: u128,
+    written: u64,
     value: Vec<u8>,
 }
 
@@ -20,10 +20,10 @@ impl Generator {
     }
 
     pub fn next(&mut self) -> Option<(Vec<u8>, Vec<u8>)> {
-        if self.written < self.config.limit as u128 * 1_000_000 {
+        if self.written < self.config.limit as u64 * 1_000_000 {
             let key = self.written.to_be_bytes().to_vec();
             let data = self.value.clone();
-            self.written += (key.len() + data.len()) as u128;
+            self.written += (key.len() + data.len()) as u64;
             Some((key, data))
         } else {
             None
